% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericmind_any_medication.r
\name{any_medication}
\alias{any_medication}
\title{3rd function: any_medication()
syntax: any_medication(dataset, new,  summ_by, summ_of)
dataset: It is the whole imported dataset.
new:     It is the filtered dataset(a subset of whole dataset) that you want to summarize.
summ_by: It is the variable that you use to summarize the table: " "
summ_of: It is the collection of variables that you want to summarize: c(" ", " ", )
Important notes:
For any_medication(): make sure the table has a unique variable 'USUBJID'}
\usage{
any_medication(dataset, new, summ_by, summ_of)
}
\description{
3rd function: any_medication()
syntax: any_medication(dataset, new,  summ_by, summ_of)
dataset: It is the whole imported dataset.
new:     It is the filtered dataset(a subset of whole dataset) that you want to summarize.
summ_by: It is the variable that you use to summarize the table: " "
summ_of: It is the collection of variables that you want to summarize: c(" ", " ", )
Important notes:
For any_medication(): make sure the table has a unique variable 'USUBJID'
}
