% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_anymedication.R
\name{nm_anymedication}
\alias{nm_anymedication}
\title{Any medication package is used for creating summary tables of table 1.25 and 1.26 specifically.
First we have to import the dataset we want to work on.
Then we have to filter out the values that we want to summarize.}
\usage{
nm_anymedication(dataset, new, summ_by, summ_of)
}
\description{
Syntax:
nm_anymedication(dataset, new, summ_by, summ_of)
}
\details{
dataset: Imported dataset
new:     Filtered dataset that we want to summarize
summ_by: Varible on the basis of which we want to summarize
summ_of: Variable that we want to summarize.
This package is used to generate count and count percentage
}
