% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_stratify.R
\name{nm_stratify}
\alias{nm_stratify}
\title{stratifying function:
for stratify
remember that while stratifying the summarized table, the values in the columns of summarized table that we are stratifying should occur in equal frequencies. i.e values in column a must occur same as the adjacent values in column b.
syntax:
nm_stratify<- function(unsummarized_table = , summarized_table = , stratify_by = " ", stratify= c(" ", " ",))}
\usage{
nm_stratify(unsummarized_table, summarized_table, stratify_by, stratify)
}
\description{
unsummarized_table: It is the table that is used for creating a normal summary table.
summarized_table:   It is the summary table created by help of unsummarized_table using nm_table() function.
stratify_by:        It is the varibale by which you want to categorize.
stratify:           It is the collection of variables that you want to stratify: c(" ", " ")
}
