% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_table.R
\name{nm_table}
\alias{nm_table}
\title{\enumerate{
\item Normal Summary Table
syntax: nm_table(dataset = , summ_of= c(" ", " ", ), summ_by= " "  , Count_cat= 'Y' or 'N',  Counts= 'Y' or 'N', Mean='Y' or 'N', Median='Y' or 'N', SD='Y' or 'N', Min_val='Y' or 'N', Max_val='Y' or 'N')
remember to omit the NA values while working with continuous variables.
dataset:   It is the dataset that we want to summarize.
summ_of:   It is the collection of variables that we want to summarize.
summ_by:   It is the variable by which we summarize.
Count_cat: It returns the summary of total numbers for categorical variables:'Y' turns it on and 'N' or any other character turns it off.
Counts:    It returns the summary of total numbers for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Mean:      It returns the mean for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Median:    It returns the median for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
SD:        It returns the standard deviation for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Min_val:   It returns the minimum value for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Max_val:   It returns the Maximum value for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
@export
}}
\usage{
nm_table(
  dataset,
  summ_of,
  summ_by,
  Count_cat,
  Counts,
  Mean,
  Median,
  SD,
  Min_val,
  Max_val
)
}
\description{
\enumerate{
\item Normal Summary Table
syntax: nm_table(dataset = , summ_of= c(" ", " ", ), summ_by= " "  , Count_cat= 'Y' or 'N',  Counts= 'Y' or 'N', Mean='Y' or 'N', Median='Y' or 'N', SD='Y' or 'N', Min_val='Y' or 'N', Max_val='Y' or 'N')
remember to omit the NA values while working with continuous variables.
dataset:   It is the dataset that we want to summarize.
summ_of:   It is the collection of variables that we want to summarize.
summ_by:   It is the variable by which we summarize.
Count_cat: It returns the summary of total numbers for categorical variables:'Y' turns it on and 'N' or any other character turns it off.
Counts:    It returns the summary of total numbers for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Mean:      It returns the mean for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Median:    It returns the median for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
SD:        It returns the standard deviation for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Min_val:   It returns the minimum value for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
Max_val:   It returns the Maximum value for continuous variables: 'Y' turns it on and 'N' or any other character turns it off.
@export
}
}
