% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numericmind_stratify_summary.r
\name{nm_stratify}
\alias{nm_stratify}
\title{2nd function: nm_stratify()
syntax:nm_stratify(unsummarized_table, summarized_table, stratify_by, stratify)
unsummarized_table: It is the table that you used for creating a normal summary table by nm_table() function.
summarized_table:   It is the table that you get by applying nm_table() function.
stratify_by:        It is the variable that is used for categorizing the summary table. : " "
stratify:           It is the collection of variables that are grouped under the stratify_by variable:  c(" ", " ")
Important notes:
For nm_stratify(): make sure that the variables that are categorized have same frequency (i.e elements in variable'a' occurs in same frequency as elements in variable 'b')}
\usage{
nm_stratify(unsummarized_table, summarized_table, stratify_by, stratify)
}
\description{
2nd function: nm_stratify()
syntax:nm_stratify(unsummarized_table, summarized_table, stratify_by, stratify)
unsummarized_table: It is the table that you used for creating a normal summary table by nm_table() function.
summarized_table:   It is the table that you get by applying nm_table() function.
stratify_by:        It is the variable that is used for categorizing the summary table. : " "
stratify:           It is the collection of variables that are grouped under the stratify_by variable:  c(" ", " ")
Important notes:
For nm_stratify(): make sure that the variables that are categorized have same frequency (i.e elements in variable'a' occurs in same frequency as elements in variable 'b')
}
