% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Numericmind.R, R/numericmind_table_simple.r
\name{nm_table}
\alias{nm_table}
\title{2nd function: nm_stratify()
syntax:nm_stratify(unsummarized_table, summarized_table, stratify_by, stratify)
unsummarized_table: It is the table that you used for creating a normal summary table by nm_table() function.
summarized_table:   It is the table that you get by applying nm_table() function.
stratify_by:        It is the variable that is used for categorizing the summary table. : " "
stratify:           It is the collection of variables that are grouped under the stratify_by variable:  c(" ", " ")
3rd function: any_medication()
syntax: any_medication(dataset, new,  summ_by, summ_of)
dataset: It is the whole imported dataset.
new:     It is the filtered dataset(a subset of whole dataset) that you want to summarize.
summ_by: It is the variable that you use to summarize the table: " "
summ_of: It is the collection of variables that you want to summarize: c(" ", " ", )
Important notes:
For nm_table(), if you are creating summary table for a numeric type variable(continuous variable),
make sure to omit the NA values before using the function. If the table only consists of categorical variables, there is no need to omit the NA values.
For nm_stratify(): make sure that the variables that are categorized have same frequency (i.e elements in variable'a' occurs in same frequency as elements in variable 'b')}
\usage{
nm_table(
  dataset,
  summ_of,
  summ_by,
  Count_cat,
  Counts,
  Mean,
  Median,
  SD,
  Min_val,
  Max_val
)

nm_table(
  dataset,
  summ_of,
  summ_by,
  Count_cat,
  Counts,
  Mean,
  Median,
  SD,
  Min_val,
  Max_val
)
}
\description{
For any_medication(): make sure the table has a unique variable 'USUBJID'
}
